# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SmartFin API",
    "version": "0.1.0",
    "description": "Personal and family finance API. Amounts are decimal strings; instants are ISO 8601 UTC; financial dates are YYYY-MM-DD."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sf_session",
        "description": "Session cookie set by sign-in (named __Host-sf_session over HTTPS). State-changing requests also need the X-SmartFin-CSRF header."
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session token for the Android app (sign in with X-SmartFin-Client: android)."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "bad_request",
                  "validation_failed",
                  "unauthenticated",
                  "forbidden",
                  "csrf_failed",
                  "not_found",
                  "conflict",
                  "rate_limited",
                  "internal_error",
                  "service_unavailable"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {},
              "requestId": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ],
        "description": "Every API error uses this shape."
      },
...
      "IncomeSource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "salary",
              "business",
              "freelance",
              "rent",
              "interest",
              "dividend",
              "bonus",
              "pension",
              "other"
            ]
          },
          "expectedAmount": {
            "type": [
              "string",
              "null"
            ],
            "pattern": "^-?\\d{1,16}(\\.\\d{1,2})?$",
            "description": "Decimal amount as a string (never a float)",
            "example": "123456.78"
          },
          "frequency": {
            "type": "string",
            "enum": [
              "monthly",
              "quarterly",
              "yearly",
              "irregular"
            ]
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "description": "Calendar date",
            "example": "2026-10-04"
          },
          "expectedDay": {
            "type": [
              "integer",
              "null"
            ]
          },
          "receivingAccountId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "taxNotes": {
            "type": [
              "string",
              "null"
            ]
          },
          "notes": {
            "type": [
              "string",
              "null"
            ]
          },
          "active": {
            "type": "boolean"
          },
          "isDemo": {
            "type": "boolean"
          },
          "version": {
            "type": "integer"
          },
          "period": {
            "type": "object",
            "properties": {
              "month": {
                "type": "string",
                "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
                "example": "2026-10"
              },
              "expected": {
                "type": [
                  "string",
                  "null"
                ],
                "pattern": "^-?\\d{1,16}(\\.\\d{1,2})?$",
                "description": "Null when nothing is expected this month",
                "example": "123456.78"
              },
              "received": {
                "type": "string",
                "pattern": "^-?\\d{1,16}(\\.\\d{1,2})?$",
                "description": "Income transactions linked to this source",
                "example": "123456.78"
              },
              "pending": {
                "type": [
                  "string",
                  "null"
                ],
                "pattern": "^-?\\d{1,16}(\\.\\d{1,2})?$",
                "description": "Decimal amount as a string (never a float)",
                "example": "123456.78"
              },
              "status": {
                "type": "string",
                "enum": [
                  "received",
                  "partial",
                  "pending",
                  "not_expected"
                ]
              }
            },
            "required": [
              "month",
              "expected",
              "received",
              "pending",
              "status"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "kind",
          "expectedAmount",
          "frequency",
          "startDate",
          "expectedDay",
          "receivingAccountId",
          "taxNotes",
          "notes",
          "active",
          "isDemo",
          "version",
          "period"
        ]
      },
      "IncomeSourceInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60
          },
          "kind": {
            "type": "string",
            "enum": [
              "salary",
              "business",
              "freelance",
              "rent",
              "interest",
              "dividend",
              "bonus",
              "pension",
              "other"
            ]
          },
          "expectedAmount": {
            "type": [
              "string",
              "null"
            ],
            "pattern": "^-?\\d{1,16}(\\.\\d{1,2})?$",
            "description": "Positive decimal amount as a string",
            "example": "1499.00"
          },
          "frequency": {
            "type": "string",
            "enum": [
              "monthly",
              "quarterly",
              "yearly",
              "irregular"
            ]
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "description": "Calendar date",
            "example": "2026-10-04"
          },
          "expectedDay": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 1,
            "maximum": 31
          },
          "receivingAccountId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "taxNotes": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 300
          },
          "notes": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 500
          },
          "active": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "kind",
          "frequency",
          "startDate"
        ]
      },
      "IncomeSourceUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60
          },
          "kind": {
            "type": "string",
            "enum": [
              "salary",
              "business",
              "freelance",
              "rent",
              "interest",
              "dividend",
              "bonus",
              "pension",
              "other"
            ]
          },
          "expectedAmount": {
            "type": [
              "string",
              "null"
            ],
            "pattern": "^-?\\d{1,16}(\\.\\d{1,2})?$",
            "description": "Positive decimal amount as a string",
            "example": "1499.00"
          },
          "frequency": {
            "type": "string",
            "enum": [
              "monthly",
              "quarterly",
              "yearly",
              "irregular"
            ]
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "description": "Calendar date",
            "example": "2026-10-04"
          },
          "expectedDay": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 1,
            "maximum": 31
          },
          "receivingAccountId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "taxNotes": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 300
          },
          "notes": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 500
          },
          "active": {
            "type": "boolean"
          },
          "version": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "description": "Version read by the client; the update fails with 409 if it changed."
          }
        },
        "required": [
          "version"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/income-sources": {
      "get": {
        "tags": [
          "Income"
        ],
        "summary": "Income sources with expected vs received for a month",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
              "example": "2026-10"
            },
            "required": false,
            "name": "month",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IncomeSource"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Income"
        ],
        "summary": "Create an income source",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncomeSourceInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncomeSource"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid CSRF token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/income-sources/{id}": {
      "patch": {
        "tags": [
          "Income"
        ],
        "summary": "Update an income source",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncomeSourceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncomeSource"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid CSRF token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found (or not yours)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Income"
        ],
        "summary": "Delete an income source",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Success, no content"
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid CSRF token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found (or not yours)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}